Improve cache layer - Move to new folder - Make interface and abstract base with generics - Add implementation for PSObject@
